requirements_test.txt
coverage==4.4.1
flake8>=2.1.0
tox>=1.7.0
codecov>=2.0.0
pytest==4.3.0
pytest-django==3.4.7
pytest-cov==2.6.1
pytest-mock==1.10.1
